requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.28.0
python-dotenv>=0.19.0
httpx>=0.24.0

# LangChain & AI
langchain>=0.0.300
langchain-openai>=0.0.2
openai>=1.3.0

# API Integrations
tavily-python>=0.1.9
wikipedia>=1.4.0

# Security & Monitoring
python-jose>=3.3.0
passlib>=1.7.4
prometheus-fastapi-instrumentator>=6.1.0
slowapi>=0.1.8

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0 